Fixed object loading

+++ src/D3D12PathTracer/src/shaders/RayTracingHlslCompat.h
@@ -18,7 +18,7 @@
 using namespace DirectX;
 
 // Shader will use byte encoding to access indices.
-typedef UINT16 Index;
+typedef UINT32 Index;
 #endif
 
 struct SceneConstantBuffer
